refresh_directories.sh: abort if no dirs exist; clean up after

+++ shell_scripts/refresh_directories.sh
@@ -6,14 +6,14 @@
 
 TEMPNAME=__REFRESH_TEMP__
 
-echo "Refreshing all directories in this directory"
-
 rmdir -p $TEMPNAME 2>/dev/null
 test -e $TEMPNAME && \
 	echo "$TEMPNAME dir exists after rmdir attempt; aborting" && exit 1
 
 for X in */
-	do echo "$X"
+	do
+	test "$X" = "*/" && echo "No directories found to refresh" && exit 1
+	echo "$X"
 	test -h "$X" && echo "Skipping symlink: $X" && continue
 	rmdir -p $TEMPNAME 2>/dev/null
 	DT="$(stat -c '%y' "$X" | cut -d. -f1)"
@@ -27,4 +27,5 @@ for X in */
 	touch -d "$DT" "$X"
 	chmod $PERM "$X"
 	chown $OWN "$X"
+	rmdir -f "$TEMPNAME" 2>/dev/null >/dev/null
 done
